% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_vol_gendata.R
\name{clus_vol_gendata}
\alias{clus_vol_gendata}
\title{Generating observations from the multiple time series with clustered volatility (ARCH-type)}
\usage{
clus_vol_gendata(
  Nlength,
  clusizes,
  phi,
  rand.mean,
  rand.sd,
  vol.par0,
  vol.par1,
  nburn = 500,
  seed = 1
)
}
\arguments{
\item{Nlength}{A numeric scalar specifying the length of each time series.}

\item{clusizes}{A vector of length \code{k} specifying the number of time series in each of the \code{k} clusters.}

\item{phi}{A numeric scalar in (-1,1) specifying the autoregressive parameter shared by all time series.}

\item{rand.mean}{A vector of length \code{N} specifying the random effects means for all \code{N} time series.}

\item{rand.sd}{A vector of length \code{N} specifying the random effects standard deviations for all \code{N} time series.}

\item{vol.par0}{A vector of length \code{k} specifying the ARCH volatility parameters (intercept) of each cluster.}

\item{vol.par1}{A vector of length \code{k} specifying the ARCH volatility parameters (slope) of each cluster.}

\item{nburn}{A numeric scalar specifying the number of timepoints to discard from simulation.}

\item{seed}{A numeric scalar specifying a seed number for reproducibility.}
}
\value{
A dataframe that includes all simulated observations \code{Y(t)}, \code{Y(t-1)}, cluster and series IDs.
Variables included are \code{clus}, \code{sid}, \code{t}, \code{y}, and \code{y1}.
}
\description{
Generating observations from the multiple time series with clustered volatility (ARCH-type)
}
\examples{
set.seed(1)
Mydata <- clus_vol_gendata(Nlength = 100,clusizes = c(5,5),phi = 0.6,rand.mean = rep(0,10),
                           rand.sd = rep(1,10),vol.par0 = c(1,1),vol.par1 = c(0,0),
                           nburn = 500,seed = 1)

}
