% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_vol_estimate.R
\name{nonpar_vol_estimate}
\alias{nonpar_vol_estimate}
\title{Estimating the multiple time series with clustered volatility (ARCH-type) model through the backfitting algorithm}
\usage{
nonpar_vol_estimate(data, max.iter = 100, cut = 1e-05)
}
\arguments{
\item{data}{A dataframe that includes a column of cluster IDs, series IDs, observations \code{Y(t)}, and lagged values \code{Y(t-1)}.
Variables required are \code{clus}, \code{sid}, \code{y}, and \code{y1}.}

\item{max.iter}{A positive integer specifying the number of maximum allowable iteration before force stopping the algorithm.}

\item{cut}{A numeric scalar (close to zero) specifying the convergence criterion cut-off (absolute percent change).}
}
\value{
A list containing \code{num_iter}, \code{ar}, \code{reff}, and \code{vol_par}. These objects are required for the \code{nonpar_vol_test} function.
}
\description{
Estimating the multiple time series with clustered volatility (ARCH-type) model through the backfitting algorithm
}
\examples{
set.seed(1)
Mydata <- clus_vol_gendata(Nlength = 100,clusizes = c(5,5),phi = 0.6,rand.mean = rep(0,10),
                           rand.sd = rep(1,10),vol.par0 = c(1,1),vol.par1 = c(0,0.4),
                           nburn = 500,seed = 1)
nonpar_vol_estimate(Mydata,max.iter=100,cut=0.00001)

}
