% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_vol_test.R
\name{nonpar_vol_test}
\alias{nonpar_vol_test}
\title{Nonparametric test for clustered volatility in multiple time series model based on bootstrap resampling scheme}
\usage{
nonpar_vol_test(
  data,
  alpha = 0.05,
  B = 100,
  max.iter = 100,
  cut = 1e-05,
  seed = 1
)
}
\arguments{
\item{data}{A dataframe that includes a column of cluster IDs, series IDs, observations \code{Y(t)}, and lagged values \code{Y(t-1)}.
Variables required are \code{clus}, \code{sid}, \code{y}, and \code{y1}.}

\item{alpha}{A numeric scalar in (0,1) specifying the level of significance for the bootstrap confidence interval.}

\item{B}{A positive integer specifying the number of bootstrap replicates for computing the confidence intervals.}

\item{max.iter}{A positive integer specifying the number of maximum allowable iteration before force stopping the algorithm.}

\item{cut}{A numeric scalar (close to zero) specifying the convergence criterion cut-off (absolute percent change).}

\item{seed}{A numeric scalar specifying a seed number for reproducibility.}
}
\value{
A list containing \verb{Number of Iterations}, \verb{Estimated AR coefficient}, \verb{Predicted Series Random Effects},
\verb{Estimated ARCH(1) Slope Parameters} and \verb{Bootstrap CI for Slope Parameters}.
}
\description{
Nonparametric test for clustered volatility in multiple time series model based on bootstrap resampling scheme
}
\examples{
set.seed(1)
Mydata <- clus_vol_gendata(Nlength = 100,clusizes = c(5,5),phi = 0.6,rand.mean = rep(0,10),
                           rand.sd = rep(1,10),vol.par0 = c(1,1),vol.par1 = c(0,0),
                           nburn = 500,seed = 1)
nonpar_vol_test(Mydata,B=100,max.iter=100,cut=0.00001)

set.seed(2)
Mydata <- clus_vol_gendata(Nlength = 100,clusizes = c(5,5),phi = 0.6,rand.mean = rep(0,10),
                           rand.sd = rep(1,10),vol.par0 = c(1,1),vol.par1 = c(0,0.4),
                           nburn = 500,seed = 1)
nonpar_vol_test(Mydata,B=100,max.iter=100,cut=0.00001)

set.seed(3)
Mydata <- clus_vol_gendata(Nlength = 100,clusizes = c(5,5),phi = 0.6,rand.mean = rep(0,10),
                           rand.sd = rep(1,10),vol.par0 = c(1,1),vol.par1 = c(0.4,0.4),
                           nburn = 500,seed = 1)
nonpar_vol_test(Mydata,B=100,max.iter=100,cut=0.00001)

}
