% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonparVolTest.R
\docType{package}
\name{NonparVolTest}
\alias{NonparVolTest}
\title{NonparVolTest: A package for simulating observations, estimating parameters and nonparametric testing for clustered volatility.}
\description{
The \code{NonparVolTest} package provides three important functions, namely,
\code{clus_vol_gendata()}, \code{nonpar_vol_estimate()} and \code{nonpar_vol_test()}.
}
\section{\code{NonparVolTest} functions}{

The \code{clus_vol_gendata()} function generates observations from the multiple time series with clustered volatility (ARCH-type).
The \code{nonpar_vol_estimate()} function estimates the parameters of the multiple time series with clustered volatility (ARCH-type) model
through the backfitting algorithm.
The \code{nonpar_vol_test()} function performs the nonparametric test for clustered volatility in multiple time series model
based on bootstrap resampling scheme.
}

